requirements.txt
python-dotenv>=1.0.0
numpy==2.2.5
scipy==1.15.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0
pydantic==2.8.2
pandas==3.0.1
joblib==1.4.2
lightgbm==4.6.0
scikit-learn==1.8.0
h3==4.1.0
ortools==9.15.6755 # 9.15.6755 (Checked)
